requirements/dev.txt
ipdb
django-debug-toolbar
django-silk